+++ src/commonMain/kotlin/io/github/koalaplot/core/xygraph/LinearAxisModel.kt
@@ -15,39 +15,6 @@ public interface ILinearAxisModel<T> : AxisModel<T> where T : Comparable<T>, T :
     public val minimumMajorTickSpacing: Dp
 }
 
-/**
- * Create and remember a LinearAxisModel.
- */
-@Deprecated(message = "Use rememberFloatLinearAxisModel", replaceWith = ReplaceWith("rememberFloatLinearAxisModel"))
-@Composable
-public fun rememberLinearAxisModel(
-    range: ClosedFloatingPointRange<Float>,
-    zoomRangeLimit: Float = (range.endInclusive - range.start) * ZoomRangeLimitDefault.toFloat(),
-    minimumMajorTickIncrement: Float = (range.endInclusive - range.start) * MinimumMajorTickIncrementDefault,
-    minimumMajorTickSpacing: Dp = 50.dp,
-    minorTickCount: Int = 4,
-    allowZooming: Boolean = true,
-    allowPanning: Boolean = true,
-): FloatLinearAxisModel = remember(
-    range,
-    zoomRangeLimit,
-    minimumMajorTickIncrement,
-    minimumMajorTickSpacing,
-    minorTickCount,
-    allowZooming,
-    allowPanning
-) {
-    FloatLinearAxisModel(
-        range,
-        zoomRangeLimit,
-        minimumMajorTickIncrement,
-        minimumMajorTickSpacing,
-        minorTickCount,
-        allowZooming,
-        allowPanning
-    )
-}
-
 internal fun <X, Y> List<Point<X, Y>>.toXList(): List<X> = object : AbstractList<X>() {
     override val size: Int
         get() = this@toXList.size